+++ goldens/circular-deps/packages.json
@@ -34,62 +34,8 @@
     "packages/core/src/change_detection/differs/keyvalue_differs.ts"
   ],
   [
-    "packages/core/src/di/create_injector.ts",
-    "packages/core/src/di/injector.ts"
-  ],
-  [
-    "packages/core/src/di/create_injector.ts",
-    "packages/core/src/di/provider_collection.ts",
-    "packages/core/src/render3/definition.ts",
-    "packages/core/src/render3/interfaces/definition.ts",
-    "packages/core/src/di/r3_injector.ts",
-    "packages/core/src/di/injector_compatibility.ts",
-    "packages/core/src/di/injector.ts"
-  ],
-  [
-    "packages/core/src/di/create_injector.ts",
-    "packages/core/src/di/provider_collection.ts",
-    "packages/core/src/render3/definition.ts",
-    "packages/core/src/render3/interfaces/definition.ts",
-    "packages/core/src/di/r3_injector.ts",
-    "packages/core/src/di/injector_token.ts",
-    "packages/core/src/di/injector.ts"
-  ],
-  [
-    "packages/core/src/di/create_injector.ts",
-    "packages/core/src/di/provider_collection.ts",
-    "packages/core/src/render3/definition.ts",
-    "packages/core/src/render3/interfaces/definition.ts",
-    "packages/core/src/di/r3_injector.ts",
-    "packages/core/src/di/injector.ts"
-  ],
-  [
-    "packages/core/src/di/create_injector.ts",
-    "packages/core/src/di/provider_collection.ts",
-    "packages/core/src/render3/definition.ts",
-    "packages/core/src/render3/interfaces/definition.ts",
-    "packages/core/src/di/r3_injector.ts",
-    "packages/core/src/di/null_injector.ts",
-    "packages/core/src/di/injector.ts"
-  ],
-  [
-    "packages/core/src/di/create_injector.ts",
-    "packages/core/src/di/provider_collection.ts",
-    "packages/core/src/render3/definition.ts",
-    "packages/core/src/render3/interfaces/definition.ts",
-    "packages/core/src/render3/interfaces/node.ts",
-    "packages/core/src/render3/interfaces/view.ts",
-    "packages/core/src/di/injector.ts"
-  ],
-  [
     "packages/core/src/di/injectable.ts",
     "packages/core/src/di/jit/injectable.ts"
-  ],
-  [
-    "packages/core/src/di/provider_collection.ts",
-    "packages/core/src/render3/definition.ts",
-    "packages/core/src/render3/interfaces/definition.ts",
-    "packages/core/src/di/r3_injector.ts"
   ],
   [
     "packages/core/src/linker/component_factory_resolver.ts",
@@ -120,26 +66,7 @@
   ],
   [
     "packages/core/src/render3/interfaces/container.ts",
-    "packages/core/src/render3/interfaces/node.ts",
     "packages/core/src/render3/interfaces/view.ts"
-  ],
-  [
-    "packages/core/src/render3/interfaces/container.ts",
-    "packages/core/src/render3/interfaces/view.ts"
-  ],
-  [
-    "packages/core/src/render3/interfaces/definition.ts",
-    "packages/core/src/render3/interfaces/node.ts",
-    "packages/core/src/render3/interfaces/view.ts"
-  ],
-  [
-    "packages/core/src/render3/interfaces/node.ts",
-    "packages/core/src/render3/interfaces/view.ts"
-  ],
-  [
-    "packages/core/src/render3/interfaces/node.ts",
-    "packages/core/src/render3/interfaces/view.ts",
-    "packages/core/src/render3/interfaces/query.ts"
   ],
   [
     "packages/core/src/render3/interfaces/query.ts",
