+++ packages/core/src/di/create_injector.ts
@@ -9,8 +9,8 @@
 import {EMPTY_ARRAY} from '../util/empty';
 import {stringify} from '../util/stringify';
 
-import {Injector} from './injector';
-import {Provider, StaticProvider} from './interface/provider';
+import type {Injector} from './injector';
+import type {Provider, StaticProvider} from './interface/provider';
 import {importProvidersFrom} from './provider_collection';
 import {getNullInjector, R3Injector} from './r3_injector';
 import {InjectorScope} from './scope';
