+++ packages/core/src/render3/component_ref.ts
@@ -37,7 +37,8 @@ import {reportUnknownPropertyError} from './instructions/element_validation';
 import {markViewDirty} from './instructions/mark_view_dirty';
 import {renderView} from './instructions/render';
 import {addToViewTree, createLView, createTView, executeContentQueries, getOrCreateComponentTView, getOrCreateTNode, initializeDirectives, invokeDirectivesHostBindings, locateHostElement, markAsComponentHost, setInputsForProperty} from './instructions/shared';
-import {ComponentDef, DirectiveDef, HostDirectiveDefs, InputFlags} from './interfaces/definition';
+import {ComponentDef, DirectiveDef, HostDirectiveDefs} from './interfaces/definition';
+import {InputFlags} from './interfaces/input_flags';
 import {NodeInputBindings, TContainerNode, TElementContainerNode, TElementNode, TNode, TNodeType} from './interfaces/node';
 import {Renderer} from './interfaces/renderer';
 import {RElement, RNode} from './interfaces/renderer_dom';
