+++ packages/core/src/render3/features/ng_onchanges_feature.ts
@@ -12,7 +12,7 @@ import {SimpleChange, SimpleChanges} from '../../interface/simple_change';
 import {assertString} from '../../util/assert';
 import {EMPTY_OBJ} from '../../util/empty';
 import {applyValueToInputField} from '../apply_value_input_field';
-import {DirectiveDef, DirectiveDefFeature, InputFlags} from '../interfaces/definition';
+import {DirectiveDef, DirectiveDefFeature} from '../interfaces/definition';
 
 /**
  * The NgOnChangesFeature decorates a component with support for the ngOnChanges
