+++ packages/core/test/render3/styling_next/static_styling_spec.ts
@@ -7,7 +7,8 @@
  */
 
 import {createTNode} from '@angular/core/src/render3/instructions/shared';
-import {AttributeMarker, TAttributes, TNode, TNodeType} from '@angular/core/src/render3/interfaces/node';
+import {AttributeMarker} from '@angular/core/src/render3/interfaces/attribute_marker';
+import {TAttributes, TNode, TNodeType} from '@angular/core/src/render3/interfaces/node';
 import {LView} from '@angular/core/src/render3/interfaces/view';
 import {enterView} from '@angular/core/src/render3/state';
 import {computeStaticStyling} from '@angular/core/src/render3/styling/static_styling';
